made search energies subclass of searchenergy, this allows them to have state in a more natural way.
