+++ include/EGPlanner/energy/contactEnergy.h
@@ -1,0 +1,14 @@
+
+#ifndef _contactenergy_h_
+#define _contactenergy_h_
+
+#include "include/EGPlanner/energy/searchEnergy.h"
+
+class ContactEnergy: public SearchEnergy
+{
+public:
+    double energy() const;
+};
+
+
+#endif
