+++ include/EGPlanner/energy/guidedAutoGraspEnergy.h
@@ -1,0 +1,14 @@
+#ifndef _guidedautograspenergy_h_
+#define _guidedautograspenergy_h_
+
+#include "include/EGPlanner/energy/searchEnergy.h"
+
+class GuidedAutoGraspQualityEnergy: public SearchEnergy
+{
+public:
+    double energy() const;
+
+};
+
+
+#endif
