+++ include/EGPlanner/energy/potentialQualityEnergy.h
@@ -1,0 +1,19 @@
+
+
+#ifndef _potentialenergy_h_
+#define _potentialenergy_h_
+
+#include "include/EGPlanner/energy/searchEnergy.h"
+
+class PotentialQualityEnergy: public SearchEnergy
+{
+public:
+    double energy() const;
+
+protected:
+    double potentialQualityScalingFunction(double dist, double cosTheta)const ;
+
+};
+
+
+#endif
