+++ include/EGPlanner/search.h
@@ -5,7 +5,7 @@
 
 enum SearchEnergyType{ENERGY_CONTACT, ENERGY_POTENTIAL_QUALITY, ENERGY_CONTACT_QUALITY, 
 					  ENERGY_AUTOGRASP_QUALITY, ENERGY_GUIDED_AUTOGRASP, ENERGY_STRICT_AUTOGRASP,
-					  ENERGY_COMPLIANT, ENERGY_DYNAMIC};
+                      ENERGY_COMPLIANT, ENERGY_DYNAMIC, ENERGY_CLOSURE};
 enum StateType{SPACE_COMPLETE, SPACE_AXIS_ANGLE, SPACE_ELLIPSOID, SPACE_APPROACH, 
 			   POSE_EIGEN, POSE_DOF, ATTRIBUTES};
 enum PlannerType{PLANNER_SIM_ANN, PLANNER_MT, PLANNER_LOOP, PLANNER_ONLINE, PLANNER_GT,
