+++ src/EGPlanner/guidedPlanner.cpp
@@ -28,6 +28,7 @@
 #include "guidedPlanner.h"
 #include "searchState.h"
 #include "searchEnergy.h"
+#include "energy/closureSearchEnergy.h"
 #include "simAnn.h"
 
 //#define GRASPITDBG
@@ -37,8 +38,7 @@ GuidedPlanner::GuidedPlanner(Hand *h)
 {
 	mHand = h;
 	init();
-	mEnergyCalculator = new ClosureSearchEnergy();
-	mEnergyCalculator->setType(ENERGY_CONTACT_QUALITY);
+    mEnergyCalculator = SearchEnergy::getSearchEnergy(ENERGY_CONTACT_QUALITY);
 	((ClosureSearchEnergy*)mEnergyCalculator)->setAvoidList( &mAvoidList );
 	mSimAnn = new SimAnn();
 	mChildClones = true;
