+++ src/EGPlanner/listPlanner.cpp
@@ -40,7 +40,7 @@ ListPlanner::ListPlanner(Hand *h)
 {
 	mHand = h;
 	init();
-	mEnergyCalculator = new SearchEnergy();
+    mEnergyCalculator = SearchEnergy::getSearchEnergy(ENERGY_CONTACT);
 	mEnergyCalculator->disableRendering(false);
 }
